fix(scheduler): change cmpTimeUs return to int32_t for consistency
- cmpTimeCycles returns int32_t but cmpTimeUs returned uint32_t,
  creating inconsistent API surface
- Use explicit static_cast<int32_t> for the unsigned subtraction
  to document the expected signed result

Model: deepseek-v4-flash deepseek
Agent: Reasonix Code

+++ src/gamepad/scheduler/scheduler.cpp
@@ -250,7 +250,7 @@ FAST_CODE int32_t Scheduler::cmpTimeCycles(uint32_t t1, uint32_t t2) {
   return static_cast<int32_t>(t1 - t2);
 }
 
-FAST_CODE uint32_t Scheduler::cmpTimeUs(uint32_t t1, uint32_t t2) { return t1 - t2; }
+FAST_CODE int32_t Scheduler::cmpTimeUs(uint32_t t1, uint32_t t2) { return static_cast<int32_t>(t1 - t2); }
 
 // ============================================================================
 // Task Queue Management
